Добавление загрузки новых коммментариев
